remove files into right package + rename server->spi app->api
