--- infra/src/main/kotlin/fr/sacane/jmanager/JmanagerBackApplication.kt
@@ -1,13 +1,0 @@
-package fr.sacane.jmanager
-
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-
-@SpringBootApplication
-class JmanagerBackApplication
-
-fun main(args: Array<String>) {
-	runApplication<JmanagerBackApplication>(*args)
-}
