--- infra/src/main/kotlin/fr/sacane/jmanager/common/WebConfig.kt
@@ -1,24 +1,0 @@
-package fr.sacane.jmanager.common
-
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.servlet.config.annotation.CorsRegistry
-import org.springframework.web.servlet.config.annotation.EnableWebMvc
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-
-@Configuration
-@EnableWebMvc
-class WebConfig: WebMvcConfigurer{
-
-    @Value("\${cors.origin.host}")
-    private lateinit var corsHost: String
-
-    override fun addCorsMappings(registry: CorsRegistry) {
-        registry.addMapping("/**")
-            .allowedOrigins(corsHost)
-            .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
-            .allowedHeaders("Authorization", "X-Requested-With", "content-type")
-            .exposedHeaders("Authorization")
-    }
-}
