+++ infra/src/main/kotlin/fr/sacane/jmanager/infrastructure/api/README.md
