+++ infra/src/main/kotlin/fr/sacane/jmanager/infrastructure/api/controller/UserEndpoint.kt
@@ -1,0 +1,5 @@
+package fr.sacane.jmanager.infrastructure.api.controller
+
+class UserEndpoint {
+
+}
