+++ infra/src/main/kotlin/fr/sacane/jmanager/infrastructure/server/README.md
