+++ infra/src/main/kotlin/fr/sacane/jmanager/infrastructure/server/entity/UserResource.kt
@@ -1,0 +1,51 @@
+package fr.sacane.jmanager.infrastructure.server.entity
+
+import fr.sacane.jmanager.domain.hexadoc.DatasourceEntity
+import org.springframework.lang.Nullable
+import javax.persistence.*
+
+
+@Table(name="User")
+@Entity
+@DatasourceEntity
+class UserResource(
+    @Id
+    @Nullable
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    @Column(name = "id_user", nullable = false)
+    open var id_user: Long? = null,
+
+    @Column(unique = true)
+    var username: String? = null,
+
+    @Column
+    var password: ByteArray? = null,
+
+    @Column(unique = true, nullable = true)
+    var email: String? = null,
+
+    @OneToMany(cascade = [CascadeType.ALL], fetch = FetchType.EAGER)
+    @JoinTable(
+        name = "user_account",
+        joinColumns = [JoinColumn(name = "id_user")],
+        inverseJoinColumns = [JoinColumn(name = "idAccount")]
+    )
+    var accounts: MutableList<AccountResource>?,
+
+    @OneToMany(cascade = [CascadeType.ALL])
+    @JoinTable(
+        name="user_categories",
+        joinColumns = [JoinColumn(name="id_user")],
+        inverseJoinColumns = [JoinColumn(name="id_category")]
+    )
+    var categories: MutableList<CategoryResource>?
+){
+    constructor(
+        username: String?,
+        password: ByteArray?,
+        email: String?,
+        accounts: MutableList<AccountResource>?,
+        categories: MutableList<CategoryResource>?
+
+    ) : this(null, username, password, email, accounts, categories)
+}
