+++ infra/src/main/kotlin/fr/sacane/jmanager/infrastructure/server/repositories/SheetRepository.kt
@@ -1,0 +1,8 @@
+package fr.sacane.jmanager.infrastructure.server.repositories
+
+import fr.sacane.jmanager.infrastructure.server.entity.SheetResource
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface SheetRepository : JpaRepository<SheetResource, Long>
