--- infra/src/main/kotlin/fr/sacane/jmanager/server/repositories/SheetRepository.kt
@@ -1,8 +1,0 @@
-package fr.sacane.jmanager.server.repositories
-
-import fr.sacane.jmanager.server.entity.SheetResource
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface SheetRepository : JpaRepository<SheetResource, Long>
